Add files via upload
For deleting duplicate files, first a sorted list of all files and their respective sizes are maintained. Two files in the list are compared for duplicity only if their sizes don't vary more than 100 bytes. On comparison, if the file is found to be exactly same, one of the copies is deleted.

For merging files, first a sorted list of all text files and their respective sizes are maintained. Two files in the list are allowed to be merged as one if their sizes don't vary more than 100 bytes and the percentage of similarity between two files is more than 70%.
User is given two options : 'Automatic Merging' and 'Manual Merging'.
Automatic Merging allows application to merge two files internally.
Manual Merging prompts user at every uncommon line between the two files to make a decision to remove the line or keep it.